build: make the codebase pass the toolchain the dependency bump brings
The bump carries ruff 0.9 -> 0.16 and mypy 1.14 -> 2.3, and the repo does not
pass either as it stands: 305 lint errors, four files the formatter would
rewrite, and one type error. `select = ["ALL"]` means every rule ruff adds
arrives switched on, so a dependabot PR is also a code review.

Real faults the new rules found, all fixed:

- Three `pytest.raises(match=...)` patterns contained unescaped metacharacters.
  `match="work_sessions.clock_in_id"` matched any character where a literal dot
  was meant, so the assertion was weaker than it read; `match="time|range"` is a
  deliberate alternation and now says so with a raw string.
- Three `int(round(...))` in `charts.py`. One-argument `round` already returns
  an int.
- An unpacked `terminal` that no assertion in its test ever read.
- A `lambda p: str(p)` that is `str`.
- Twelve implicit string concatenations inside `parametrize` lists in
  `test_migrations.py` -- the shape of the missing-comma bug, and at that length
  the trailing comma really is easy to miss. The SQL is named now, so each case
  reads as the sentence it is testing.
- `click.ParamType` became generic in Click 8.5; `TypedDate` is parameterised.

Two rules are ignored, with the reason recorded beside them. `CPY001` wants a
copyright header on all 238 files, which repeats what LICENSE and this file
already say. `PLC0415` wants every import at module scope, which would undo the
deliberate laziness the five facades and `minimum_session` are built on -- and
which `tests/test_layering.py` has a helper to distinguish.

`FBT001` fires on five dismissal callbacks whose signature Textual imposes; each
carries a `noqa` naming that, rather than the rule being turned off everywhere.
The six `# ruff: noqa: TC004` directives on the lazy facades are gone: the new
ruff does not raise TC004 there, and `RUF100` correctly called them unused.

The formatter now formats Python inside Markdown, which reflows four code blocks
in `docs/`. Worth taking: a reader copying an example now gets code that passes
the repo's own gate.

Finally, pytest 8.4 surfaces unraisable `ResourceWarning`s, and the `engine`
fixture was returning its engine rather than yielding and disposing it -- 425
"unclosed database" warnings a run. Down to 39, all in tests that build an
engine of their own, and only under `--cov`.

+++ src/flexi/components/charts.py
@@ -253,14 +253,14 @@ class Burndown(Widget):
         width = max(12, self.content_size.width)
         spent = self.total - self.remaining
         track = [EMPTY] * width
-        filled = int(round(spent / self.total * width))
+        filled = round(spent / self.total * width)
         for index in range(min(filled, width)):
             track[index] = FULL
 
         marker = (
             None
             if self.pace is None
-            else min(width - 1, max(0, int(round(self.pace / self.total * width))))
+            else min(width - 1, max(0, round(self.pace / self.total * width)))
         )
         if marker is not None:
             track[marker] = "┃"
@@ -416,7 +416,7 @@ class YearHeatmap(Widget):
         if effect == timedelta():
             return glyph, self.get_component_rich_style("chart--neutral")
         share = min(1.0, abs(effect) / self.scale)
-        step = max(1, min(DIVERGING_STEPS, int(round(share * DIVERGING_STEPS))))
+        step = max(1, min(DIVERGING_STEPS, round(share * DIVERGING_STEPS)))
         arm = "surplus" if effect > timedelta() else "deficit"
         return glyph, self.get_component_rich_style(f"chart--{arm}-{step}")
